refactor: move Context class in spack.enums
