+++ lib/spack/spack/cmd/build_env.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
 import spack.cmd.common.env_utility as env_utility
-from spack.context import Context
+from spack.enums import Context
 
 description = "dump the install environment for a spec,\nor run a command in that environment"
 section = "build"
