+++ lib/spack/spack/cmd/test_env.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
 import spack.cmd.common.env_utility as env_utility
-from spack.context import Context
+from spack.enums import Context
 
 description = (
     "run a command in a spec's test environment,\nor dump its environment to screen or file"
