+++ lib/spack/spack/user_environment.py
@@ -10,7 +10,7 @@ import spack.config
 import spack.spec
 import spack.util.environment as environment
 from spack import traverse
-from spack.context import Context
+from spack.enums import Context
 
 #: Environment variable name Spack uses to track individually loaded packages
 spack_loaded_hashes_var = "SPACK_LOADED_HASHES"
